requirements.txt
flask
python-dotenv
supabase
requests
pypdf
flask-cors
reportlab
